rearranged files and minor errors
